refactor(preset): move cucumber tests from projects.feature to modules.feature
