+++ src/test/features/modules.feature
@@ -30,3 +30,14 @@ Feature: Modules
   Scenario: Should not get properties definition for unknown module
     When I get module "not-a-real-module" properties definition
     Then I should have unknown slug "not-a-real-module" error message
+
+  Scenario: Should get presets definition
+    When I get the presets definition
+    Then I should have preset names
+      | test preset one |
+      | test preset two |
+    And I should have preset modules
+      | test-module-one   |
+      | test-module-two   |
+      | test-module-three |
+      | test-module-four  |
