+++ src/test/java/tech/jhipster/lite/module/infrastructure/primary/ModulesSteps.java
@@ -1,10 +1,10 @@
 package tech.jhipster.lite.module.infrastructure.primary;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static tech.jhipster.lite.TestProjects.lastProjectFolder;
-import static tech.jhipster.lite.TestProjects.newTestFolder;
-import static tech.jhipster.lite.cucumber.rest.CucumberRestAssertions.assertThatLastResponse;
+import static org.assertj.core.api.Assertions.*;
+import static tech.jhipster.lite.TestProjects.*;
+import static tech.jhipster.lite.cucumber.rest.CucumberRestAssertions.*;
 
+import io.cucumber.java.en.And;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import java.io.IOException;
@@ -21,11 +21,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.assertj.core.api.SoftAssertions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.MediaType;
+import org.springframework.http.*;
 import tech.jhipster.lite.cucumber.rest.CucumberRestTestContext;
 import tech.jhipster.lite.module.infrastructure.secondary.git.GitTestUtil;
 
@@ -314,5 +310,20 @@ public class ModulesSteps {
   @Then("I should have unknown slug {string} error message")
   public void shouldHaveUnknownSlugErrorMessage(String slugName) {
     assertThatLastResponse().hasHttpStatus(HttpStatus.INTERNAL_SERVER_ERROR);
+  }
+
+  @When("I get the presets definition")
+  public void getPresetsDefinition() {
+    rest.exchange("/api/presets", HttpMethod.GET, new HttpEntity<>(jsonHeaders()), Void.class);
+  }
+
+  @Then("I should have preset names")
+  public void shouldHavePresetName(List<String> names) {
+    assertThatLastResponse().hasOkStatus().hasElement("$.presets.*.name").withValues(names);
+  }
+
+  @And("I should have preset modules")
+  public void shouldHavePresetModules(List<String> modules) {
+    assertThatLastResponse().hasOkStatus().hasElement("$.presets.*.modules.*.slug").withValues(modules);
   }
 }
